# ARROW-3929: [Go] improve CSV reader memory usage